+++ libretro/libretro-glue.c
@@ -1,6 +1,7 @@
 #include "libretro-core.h"
 #include "archdep.h"
 
+/* Misc */
 int sensible_strcmp(char *a, char *b)
 {
    int i;
@@ -50,6 +51,97 @@ void remove_recurse(const char *path)
 
    closedir(dir);
    archdep_rmdir(path);
+}
+
+void path_join(char* out, const char* basedir, const char* filename)
+{
+	snprintf(out, RETRO_PATH_MAX, "%s%s%s", basedir, RETRO_PATH_SEPARATOR, filename);
+}
+
+char* path_join_dup(const char* basedir, const char* filename)
+{
+    size_t dirlen = strlen(basedir);
+    size_t seplen = strlen(RETRO_PATH_SEPARATOR);
+    size_t filelen = strlen(filename);
+    char* result = (char*)malloc(dirlen + seplen + filelen + 1);
+    strcpy(result, basedir);
+    strcpy(result + dirlen, RETRO_PATH_SEPARATOR);
+    strcpy(result + dirlen + seplen, filename);
+    return result;
+}
+
+/* Note: This function returns a pointer to a substring_left of the original string.
+ * If the given string was allocated dynamically, the caller must not overwrite
+ * that pointer with the returned value, since the original pointer must be
+ * deallocated using the same allocator with which it was allocated. The return
+ * value must NOT be deallocated using free() etc. */
+char* trimwhitespace(char *str)
+{
+  char *end;
+
+  /* Trim leading space */
+  while(isspace((unsigned char)*str)) str++;
+
+  if(*str == 0) /* All spaces? */
+    return str;
+
+  /* Trim trailing space */
+  end = str + strlen(str) - 1;
+  while(end > str && isspace((unsigned char)*end)) end--;
+
+  /* Write new null terminator character */
+  end[1] = '\0';
+
+  return str;
+}
+
+/* Returns a substring of 'str' that contains the 'len' leftmost characters of 'str'. */
+char* strleft(const char* str, int len)
+{
+	char* result = calloc(len + 1, sizeof(char));
+	strncpy(result, str, len);
+	return result;
+}
+
+/* Returns a substring of 'str' that contains the 'len' rightmost characters of 'str'. */
+char* strright(const char* str, int len)
+{
+	int pos = strlen(str) - len;
+	char* result = calloc(len + 1, sizeof(char));
+	strncpy(result, str + pos, len);
+	return result;
+}
+
+/* Returns true if 'str' starts with 'start' */
+bool strstartswith(const char* str, const char* start)
+{
+	if (strlen(str) >= strlen(start))
+		if(!strncasecmp(str, start, strlen(start)))
+			return true;
+		
+	return false;
+}
+
+/* Returns true if 'str' ends with 'end' */
+bool strendswith(const char* str, const char* end)
+{
+	if (strlen(str) >= strlen(end))
+		if(!strcasecmp((char*)&str[strlen(str)-strlen(end)], end))
+			return true;
+		
+	return false;
+}
+
+/* Removes ':PRG' */
+char *path_remove_program(char *path)
+{
+   char *last = !string_is_empty(path)
+      ? (char*)strrchr(path_basename(path), ':') : NULL;
+   if (!last)
+      return NULL;
+   if (*last)
+      *last = '\0';
+   return path;
 }
 
 /* zlib */
